DDA-126: Refactor DioryActions to its own component
